pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.15",
    "aiosqlite>=0.21.0",
    "discord-py>=2.5.2",
    "google-genai>=1.29.0",
    "google-generativeai>=0.8.5",
    "pygithub>=2.7.0",
    "python-dotenv>=1.1.1",
    "wikipedia>=1.4.0",
]
